% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-text-npc.R
\name{add_text_npc}
\alias{add_text_npc}
\title{Add text to ggplot2 plot by NPC}
\usage{
add_text_npc(
  text,
  x = 0.95,
  y = 0.95,
  just = c(1, 1),
  graph_params = list(fontsize = 10)
)
}
\arguments{
\item{text}{Annotation text}

\item{x, y}{Position within plot. Defaults to 0.95.}

\item{just}{Numeric vector of length 2 that specifies justification relative
to position. Defaults to upper right.}

\item{graph_params}{List of graphical parameters passed to \code{\link[grid:gpar]{grid::gpar()}}.
Sets font size to 10 by default.}
}
\description{
\code{add_text_npc()} wraps \code{\link[ggplot2:annotation_custom]{ggplot2::annotation_custom()}} to create a text grob
at location specified by Normalized Parent Coordinates (NPC) in a ggplot2
plot.
}
