% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{calc_dist_matrix}
\alias{calc_dist_matrix}
\title{Calculate distance matrix}
\usage{
calc_dist_matrix(commun_data, diss_index = "jaccard")
}
\arguments{
\item{commun_data}{Community data.}

\item{diss_index}{Dissimilarity index. See \code{\link[vegan:vegdist]{vegan::vegdist()}} for options.}
}
\description{
\code{calc_dist_matrix()} calculates a distance matrix from community data.
}
