% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{calc_mds}
\alias{calc_mds}
\title{Calculate MDS}
\usage{
calc_mds(dist_matrix, n_dims, max_tries = 50)
}
\arguments{
\item{dist_matrix}{Distance matrix.}

\item{n_dims}{Number of dimensions.}

\item{max_tries}{Maximum number of tries.}
}
\value{
Object of class \code{metaMDS}.
}
\description{
\code{calc_mds()} performs the MDS calculation from a distance matrix.
}
