% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{get_mds_info}
\alias{get_mds_info}
\title{Get MDS information}
\usage{
get_mds_info(mds_list)
}
\arguments{
\item{mds_list}{List of objects of class \code{metaMDS}.}
}
\value{
A \code{\link[=tibble]{tibble()}} with information about each ordination.
}
\description{
Retrieves MDS information from a list of MDSs and presents them in a tabular
format.
}
\details{
\code{converg_status} column is derived from the \code{converged} part of the \code{metaMDS}
object. \code{converged} means the number of times the best version was repeated
(see
\href{https://github.com/vegandevs/vegan/blob/0ddfa4e52444de119bf5baae200eef0c05b12c23/R/print.metaMDS.R#L20-L24}{here}).
It is originally initialized at 0, so it seems that 0 does mean no solution
was reached (see the internal \code{metaMDSiter()} \href{https://github.com/vegandevs/vegan/blob/0ddfa4e52444de119bf5baae200eef0c05b12c23/R/metaMDSiter.R}{function definition}).
}
