% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{get_mds_list}
\alias{get_mds_list}
\title{Get MDS list}
\usage{
get_mds_list(dist_matrix, dims = seq(2, 8), max_tries = 50)
}
\arguments{
\item{dist_matrix}{Distance matrix.}

\item{dims}{A length 2 vector describing the range of dimensions to ordinate
over. Defaults to 2--8.}

\item{max_tries}{Maximum number of tries. Defaults to 50.}
}
\value{
Named list of objects of class \code{metaMDS}.
}
\description{
\code{get_mds_list()} calculates a series of ordination and produces a list of
\code{metaMDS} objects from a distance matrix and a range of dimensions.
}
