% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{plot_stress_dim}
\alias{plot_stress_dim}
\title{Plot stress vs. number of dimensions}
\usage{
plot_stress_dim(mds_info)
}
\arguments{
\item{mds_info}{A \code{\link[=tibble]{tibble()}} with information from each ordination run.
Output from \code{\link[=get_mds_info]{get_mds_info()}}.}
}
\value{
A ggplot2 plot.
}
\description{
\code{plot_stress_dim()} plots the stress of the iteration run for each of its
dimensions. The color of the point represents whether the run converged or
not.
}
